[#34] - A bit of refactoring to clean up files.
